remove mysterious old wait hack
I tried a long experiment disabling:

   if [[ $prog == url-handler ]]; then
       need_wait=yes
   fi

and nothing broke, so should be safe to remove.  There was an old
comment:

    # Only wait for viewer process to exit if we have to (couldn't hard
    # link)... This could lead to problems because of the 30-second pause
    # (huh? what 30s pause was I talking about?)

I guess maybe I meant a "sleep 30" which might have predated
tmp-cleaner's looped "sleep 60"?  The idea presumably being that we
need to allow ~30s for the viewer to read the temporary file before
it's safe to delete it.  However tmp-cleaner takes a more
sophisticated approach now, using fuser to check if the tmp file is
still in use.