Experiment with disabling reverse_alias for a while
DMARC policies can force Mailman to munge the From: header, in
which case they can look like:

    From: Joe Bloggs via listname <listname@listserver.com>

If there is an alias defined for <listname@listserver.com> then
enabling reverse_alias causes the %F field of $index_format to display
as the list's alias name rather than the sender's name.  So disable it
for now and see if it causes other annoyances.